+++ src/components/search-bar.tsx
@@ -4,7 +4,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useEffect, useState } from "react";
 import { Input } from "@/components/shadcn/input";
 import { Card } from "@/components/shadcn/card";
-import Loading from "./fallback/loading";
+import Loading from "./fallback/oh-loading";
 import { Avatar, AvatarFallback, AvatarImage } from "./shadcn/avatar";
 import { getUsersByNameSearch } from "@/services/user-service";
 import { useRouter } from "next/navigation";
